scripts/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
firebase-admin>=6.2.0
google-cloud-storage>=2.10.0
opencv-python>=4.8.0
numpy>=1.24.0
pytest>=7.4.0
pytest-cov>=4.1.0